move assignment chart component to a seperate file
